projects/text_generation_scores/requirements.txt
datasets==3.0.1
editdistance==0.8.1
faiss-gpu-cu12
feng_hirst_rst_parser==0.1.4
googletrans==4.0.2
lexicalrichness==0.5.1
linguaf==0.1.0
nltk==3.9.1
numba==0.60.0
numpy==2.2.2
pandas==2.2.3
python_Levenshtein==0.26.1
semantic_text_similarity==1.0.3
sentence_splitter==1.4
sentence_transformers==3.2.0
spacy==3.8.2
torch==2.4.0
tqdm==4.67.1
transformers==4.45.1
